agent/requirements.txt
web3
python-dotenv
pytest
